Moved PFR impl to separate file
